[Fix] #4 - 코드리뷰 반영 및 버그 수정 1. adjust+ 수정 2. pangesture 오류 수정 3. RxCocoa, RxSwift 라이브러리 설치

+++ Assignment/Kakako-Clone/Kakako-Clone/Global/Extensions/adjust+.swift
@@ -16,17 +16,17 @@ extension CGFloat {
 }
 
 extension Int {
-    var adjusted: Int {
-        let ratio: Int = Int(UIScreen.main.bounds.width) / 375
-        let ratioH: Int = Int(UIScreen.main.bounds.height) / 667
-        return ratio <= ratioH ? self * ratio : self * ratioH
+    var adjusted: CGFloat {
+        let ratio: CGFloat = (UIScreen.main.bounds.width) / 375
+        let ratioH: CGFloat = (UIScreen.main.bounds.height) / 667
+        return ratio <= ratioH ? CGFloat(self) * ratio : CGFloat(self) * ratioH
     }
 }
 
 extension Double {
-    var adjusted: Double {
-        let ratio: Double = Double(UIScreen.main.bounds.width) / 375
-        let ratioH: Double = Double(UIScreen.main.bounds.height) / 667
-        return ratio <= ratioH ? self * ratio : self * ratioH
+    var adjusted: CGFloat {
+        let ratio: CGFloat = (UIScreen.main.bounds.width) / 375
+        let ratioH: CGFloat = (UIScreen.main.bounds.height) / 667
+        return ratio <= ratioH ? CGFloat(self) * ratio : CGFloat(self) * ratioH
     }
 }
